@@ -1,48 +1,44 @@
 {
   "name": "windowbuilder-reports",
   "description": "Reports and image rendering for windowbuilder project",
-  "version": "2.0.237",
+  "version": "2.0.242",
   "author": "Evgeniy Malyarov <info@oknosoft.ru> (http://www.oknosoft.ru)",
   "license": "SEE LICENSE IN LICENSE.en.md",
   "main": "index.js",
   "engines": {
     "node": "^8.8",
-    "npm": "^5.5"
+    "npm": "^6.0"
   },
   "scripts": {
     "start": "node ./index",
     "debug": "node --inspect-brk ./index",
     "build": "node ./scripts/build",
     "canvas": "apt-get install libcairo2-dev libjpeg62-turbo-dev libpango1.0-dev libgif-dev build-essential g++",
-    "meta:prebuild": "node ./scripts/meta.prebuild",
-    "meta:link": "node ./scripts/meta.link"
+    "meta:link": "node ./scripts/meta.link",
+    "drawer:fetch": "node scripts/drawer.fetch.js"
   },
   "dependencies": {
-    "@koa/cors": "^2.2.1",
-    "koa": "^2.5.0",
+    "@koa/cors": "^2.2.2",
+    "canvas": "^1.6.13",
+    "koa": "^2.6.2",
     "koa-better-router": "^2.1",
-    "md5-file": "^3.2.3",
-    "metadata-abstract-ui": "2.0.16-beta.52",
-    "metadata-core": "2.0.16-beta.52",
-    "metadata-pouchdb": "2.0.16-beta.52",
-    "paper-jsdom-canvas": "^0.11.5"
+    "md5-file": "^4.0",
+    "metadata-abstract-ui": "2.0.17-beta.13",
+    "metadata-core": "2.0.17-beta.13",
+    "metadata-pouchdb": "2.0.17-beta.13",
+    "paper-jsdom-canvas": "^0.12"
   },
   "devDependencies": {
-    "eslint": "^4.18.1",
+    "eslint": "^5.9.0",
     "eslint-config-react-app": "^2.1.0",
-    "eslint-plugin-promise": "^3.6.0",
-    "eslint-plugin-react": "^7.7.0",
-    "gulp": "^3.9",
-    "gulp-concat": "^2.6",
-    "gulp-rename": "^1.2",
-    "gulp-replace": "^0.6.1",
-    "gulp-strip-comments": "^2.5.2",
-    "gulp-wrap": "^0.13.0",
-    "rollup": "^0.51",
-    "rollup-plugin-cleanup": "^2.0",
+    "eslint-plugin-promise": "^3.8.0",
+    "eslint-plugin-react": "^7.11.1",
+    "node-fetch": "^2.3.0",
+    "rollup": "^0.66.6",
+    "rollup-plugin-cleanup": "^3.0.0",
     "rollup-plugin-node-builtins": "^2.1.2",
-    "rollup-plugin-node-resolve": "^3.0.3",
-    "rollup-plugin-replace": "^2.0.0"
+    "rollup-plugin-node-resolve": "^3.4.0",
+    "rollup-plugin-replace": "^2.1.0"
   },
   "homepage": "http://www.oknosoft.ru/zd/",
   "repository": {
