@@ -3,7 +3,8 @@
 const debug = require('debug')('wb:router');
 debug('start');
 
-import builder from './windowbuilder';
+import builder from './builder';
+import search from './search';
 
 import Router from 'koa-better-router';
 const rep = Router({ prefix: '/r' });
@@ -13,6 +14,7 @@ rep.loadMethods()
     await next();
     ctx.body = `Reports: try out <a href="/r/img">/r/img</a> too`
   })
-  .get('/img/:class/:ref', builder);
+  .get('/img/:class/:ref', builder)
+  .post('/_find', search);
 
 export default rep;
